Add overflow menu support for sizing policy
